Changed the return type of count from int to long

+++ src/main/java/za/co/no9/util/EnumerationUtils.java
@@ -5,8 +5,8 @@ public class EnumerationUtils {
         return enumeration.hasMoreElements() ? Optional.of(enumeration.nextElement()) : Optional.<T>empty();
     }
 
-    public static <T> int size(FilteredEnumeration<T> enumeration) {
-        int size = 0;
+    public static <T> long size(FilteredEnumeration<T> enumeration) {
+        long size = 0;
         while (enumeration.hasMoreElements()) {
             enumeration.nextElement();
             size += 1;
